requirements.txt
torch
torchvision
torchaudio
safetensors
mlx
mlx_lm
# mlx_image
jaxtyping
transformers>=5.0.0,<6.0
huggingface_hub
hf-xet==1.2.0
peft
trl
accelerate
diffusers
sentencepiece
gradio
scipy
Pillow
numpy
regex
scikit-image
scikit-learn
tqdm
pandas
types-pyyaml
matplotlib
opencv-python
opencv-contrib-python
opencv-python-headless
pycocotools
